Code working well for most cases. All steps added

+++ src/include/kernel.hpp
@@ -9,9 +9,9 @@ extern "C" void run_WarmUp(dim3 blocks,dim3 block_size);
 /*Phase 6 start fractal XORing*/
 extern "C" void run_FracXor(uint8_t *in,uint8_t *out,uint8_t *fractal,dim3 blocks,dim3 block_size);
 /*Phase 7 start Arnold Map Encryption*/
-extern "C" void run_EncGenCatMap(uint8_t *in,uint8_t *out,uint16_t *colRotate,uint16_t *rowRotate,dim3 blocks,dim3 block_size);
+extern "C" void run_EncGenCatMap(uint8_t *in,uint8_t *out,uint32_t *colRotate,uint32_t *rowRotate,dim3 blocks,dim3 block_size);
 /*Phase 8 start Arnold Map Decryption*/
-extern "C" void run_DecGenCatMap(uint8_t *in,uint8_t *out,uint16_t *colRotate,uint16_t *rowRotate,dim3 blocks,dim3 block_size);
+extern "C" void run_DecGenCatMap(uint8_t *in,uint8_t *out,uint32_t *colRotate,uint32_t *rowRotate,dim3 blocks,dim3 block_size);
 /*Phase 9 start mapping image to Arnold Map Table*/
 extern "C" void run_ArMapTable(uint32_t *in,uint32_t *out,dim3 blocks, dim3 block_size);
 extern "C" void run_ArMapTabletoImg(uint8_t *in,uint8_t *out,uint32_t *table,dim3 blocks,dim3 block_size);
